Resave data & log compression targets in raw data. See https://r-pkgs.org/data.html#sec-data-data:~:text=updated%20as%20necessary.-,Submitting%20to%20CRAN,-Generally%2C%20package%20data
This is still very likely too big. We will either need to include less data, or pull the sample data off into gsm.sampledata or something similar (which would *only* contain the sample datasets). We're pretty close to the edge, so we can probably make it work if we make the data just a *little* smaller, but I don't think they'll be happy with it as-is (since the data is ONLY for demonstration purposes).

I also got rid of the unused "glossary", and resaved images to try to save some space.