Make tests less noisy. (#34)
* Make tests less noisy.

It will be easier to deal with CRAN issues if test logs are clean.

* Don't show expected/known messages in test logs.

DO repeat yourself in tests; if something runs outside a test_that, debugging an individual test can be difficult.

Avoid pointless comments. If it's obvious what a step does, don't add a comment to explain it.

* Resave data & log compression targets in raw data. See https://r-pkgs.org/data.html#sec-data-data:~:text=updated%20as%20necessary.-,Submitting%20to%20CRAN,-Generally%2C%20package%20data

This is still very likely too big. We will either need to include less data, or pull the sample data off into gsm.sampledata or something similar (which would *only* contain the sample datasets). We're pretty close to the edge, so we can probably make it work if we make the data just a *little* smaller, but I don't think they'll be happy with it as-is (since the data is ONLY for demonstration purposes).

I also got rid of the unused "glossary", and resaved images to try to save some space.

* Rever image that got bigger.

* Delete unused app.

* Delete glossary (reapply).

* Redocument.

* Remove remaining references to other gsm packages.

* Remove unused tidyeval reexports.

* Tweak README.

We should do a full sweep through the readme, I'm pretty sure it's out of date.

* Consistent "gsm" not "GSM".

* Minimize import footprint.

* Get rid of `%>%`-specific `.` usage to avoid R CMD Check NOTE.

* Use standard test filenames.

This way usethis::use_test() and usethis::use_r() function properly (and my ctrl-alt-t and ctrl-alt-r hotkeys allow me to hop between the R file and the corresponding test file).

* More footprint minimization.

* Re-elevate packages for RunWorkflow.